Add version to fuel raster table
