+++ api/alembic/versions/a1553ead7fde_add_processed_fuel_raster_table.py
@@ -1,0 +1,42 @@
+"""add processed fuel raster table
+
+Revision ID: a1553ead7fde
+Revises: 42a9dae10dca
+Create Date: 2025-05-26 15:42:34.221962
+
+"""
+
+from alembic import op
+import sqlalchemy as sa
+from sqlalchemy.dialects import postgresql
+from wps_shared.db.models.common import TZTimeStamp
+
+# revision identifiers, used by Alembic.
+revision = "a1553ead7fde"
+down_revision = "42a9dae10dca"
+branch_labels = None
+depends_on = None
+
+
+def upgrade():
+    # ### commands auto generated by Alembic ###
+    op.create_table(
+        "fuel_type_raster",
+        sa.Column("id", sa.Integer(), nullable=False),
+        sa.Column("year", sa.Integer(), nullable=False),
+        sa.Column("version", sa.Integer(), nullable=False),
+        sa.Column("xsize", sa.Integer(), nullable=False),
+        sa.Column("ysize", sa.Integer(), nullable=False),
+        sa.Column("object_store_path", sa.String(), nullable=False),
+        sa.Column("content_hash", sa.String(), nullable=False),
+        sa.Column("create_timestamp", TZTimeStamp(), nullable=False),
+        sa.PrimaryKeyConstraint("id"),
+        comment="Processed fuel type rasters.",
+    )
+    # ### end Alembic commands ###
+
+
+def downgrade():
+    # ### commands auto generated by Alembic ###
+    op.drop_table("fuel_type_raster")
+    # ### end Alembic commands ###
