--- api/alembic/versions/e9b263278479_add_fuel_raster_table.py
@@ -1,56 +1,0 @@
-"""add fuel raster table
-
-Revision ID: e9b263278479
-Revises: 6a31639810b0
-Create Date: 2025-04-15 14:56:17.922965
-
-"""
-
-from alembic import op
-import sqlalchemy as sa
-
-# revision identifiers, used by Alembic.
-revision = "e9b263278479"
-down_revision = "42a9dae10dca"
-branch_labels = None
-depends_on = None
-
-
-def upgrade():
-    # ### commands auto generated by Alembic ###
-    op.create_table(
-        "fuel_type_raster",
-        sa.Column("id", sa.Integer(), nullable=False),
-        sa.Column("year", sa.Integer(), nullable=False),
-        sa.Column("xsize", sa.Integer(), nullable=False),
-        sa.Column("ysize", sa.Integer(), nullable=False),
-        sa.Column("object_store_path", sa.String(), nullable=False),
-        sa.Column("content_hash", sa.String(), nullable=False),
-        sa.Column("create_timestamp", sa.TIMESTAMP(timezone=True), nullable=False),
-        sa.PrimaryKeyConstraint("id"),
-        comment="Processed fuel type rasters.",
-    )
-    op.create_index(
-        op.f("ix_fuel_type_raster_create_timestamp"),
-        "fuel_type_raster",
-        ["create_timestamp"],
-        unique=False,
-    )
-    op.create_index(
-        op.f("ix_fuel_type_raster_object_store_path"),
-        "fuel_type_raster",
-        ["object_store_path"],
-        unique=False,
-    )
-    op.create_index(op.f("ix_fuel_type_raster_year"), "fuel_type_raster", ["year"], unique=False)
-
-    # ### end Alembic commands ###
-
-
-def downgrade():
-    # ### commands auto generated by Alembic ###
-    op.drop_index(op.f("ix_fuel_type_raster_year"), table_name="fuel_type_raster")
-    op.drop_index(op.f("ix_fuel_type_raster_object_store_path"), table_name="fuel_type_raster")
-    op.drop_index(op.f("ix_fuel_type_raster_create_timestamp"), table_name="fuel_type_raster")
-    op.drop_table("fuel_type_raster")
-    # ### end Alembic commands ###
